refactored test data creation
