--- src/main/java/com/epam/report/portal/driver/DriverManager.java
@@ -1,5 +1,0 @@
-package com.epam.report.portal.driver;
-
-public class DriverManager {
-    //TODO
-}
