+++ src/main/java/com/epam/report/portal/factory/driver/DriverManager.java
@@ -1,0 +1,5 @@
+package com.epam.report.portal.factory.driver;
+
+public class DriverManager {
+    //TODO
+}
